PL-296: User Data Audit: Update Acts As Revisioned Gem
Add ActsAsRevisionable#build_revision

This builds an revision record without saving it to the database.
This is necessary if you'd like to compare the current state to a stored revision.

It also brings arbitrary attributes forward. This work was done b/c it
was just re-reading from the database, so non-db backed attributes go
swallowed.